% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompile_model_headers.R
\name{precompile_model_headers}
\alias{precompile_model_headers}
\title{Pre-compile header files used in all Stan models}
\usage{
precompile_model_headers(force = TRUE)
}
\arguments{
\item{force}{Whether to overwrite existing pre-compiled files}
}
\description{
Reduces model compilation time by pre-compiling repeated includes
}
