% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmodel.R
\name{stan_model}
\alias{stan_model}
\title{Wrapper function to source and compile Stan model}
\usage{
stan_model(model_path = NULL, model_code = NULL, include_paths = NULL, ...)
}
\arguments{
\item{model_path}{The path to a \code{.stan} file}

\item{model_code}{The stan model as a single string}

\item{include_paths}{A character vector of paths to find \code{.stan} files referenced by \verb{#include} statements}

\item{...}{Any additional arguments to be passed to the $compile() method}
}
\description{
Creates the base Stan class containing the compiled model
and access to sampling/optimisation/variational methods
}
